requirements.txt
uvicorn[standard]
PyJWT
fastapi
websockets
pydantic
sqlalchemy
requests
openai
mysql-connector-python
Pillow
alembic